Move AvatarHandler
